SIDM-6801 - Fix for cross-browser tests
