+++ package.json
@@ -15,7 +15,7 @@
     "test:coverage": "jest --coverage",
     "test:smoke": "NODE_TLS_REJECT_UNAUTHORIZED=0 jest -c jest.smoke.config.js",
     "test:crossbrowser": "./src/test/cross-browser/runCrossBrowserTests.sh",
-    "test:crossbrowser-e2e": "NODE_TLS_REJECT_UNAUTHORIZED=0 codeceptjs run-multiple --all -c ./src/test/cross-browser/saucelabs.conf.js",
+    "test:crossbrowser-e2e": "NODE_TLS_REJECT_UNAUTHORIZED=0 codeceptjs --config src/test/cross-browser/saucelabs.conf.js run",
     "test:functional:min": "NODE_TLS_REJECT_UNAUTHORIZED=0 codeceptjs --config ./src/test/functional/codecept.conf.js run-workers 8 parallel  --grep \"(?=.*)^(?!.*@Nightly)\"",
     "test:functional:all": "NODE_TLS_REJECT_UNAUTHORIZED=0 codeceptjs --config ./src/test/functional/codecept.conf.js run-workers 8 parallel",
     "test:functional": "npm-run-all -scl test:functional:min test:allure-report test:pa11y",
